Sunny/9999/requirements.txt
click==7.1.2
Flask==1.1.2
flask-paginate==0.7.0
Flask-Pagination==0.0.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.5
pandas==1.0.4
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
